Split docs truth tests
