+++ tests/docs_truth/repo_hygiene.rs
@@ -1,0 +1,126 @@
+use super::*;
+
+#[test]
+fn v1_spec_is_single_source_of_truth_and_old_spec_is_quarantined() {
+    let old_spec = read("LANGUAGE_SPEC.zen");
+    let old_spec_header = old_spec.lines().take(20).collect::<Vec<_>>().join("\n");
+
+    assert!(
+        old_spec_header.contains("Archived aspirational notes"),
+        "LANGUAGE_SPEC.zen must be explicitly quarantined at the top"
+    );
+    assert!(
+        old_spec_header.contains("docs/V1_SPEC.md is the versioned v1 source of truth"),
+        "LANGUAGE_SPEC.zen must point to docs/V1_SPEC.md as the source of truth"
+    );
+}
+
+#[test]
+fn aspirational_stdlib_is_explicitly_experimental() {
+    let stdlib_readme = read("stdlib/README.md");
+
+    for required in [
+        "experimental",
+        "not part of the implemented v1 surface",
+        "must parse, typecheck, and build",
+        "docs/V1_SPEC.md",
+    ] {
+        assert!(
+            stdlib_readme.contains(required),
+            "stdlib/README.md is missing experimental gate text: {required}"
+        );
+    }
+}
+
+#[test]
+fn ci_and_release_only_advertise_existing_targets() {
+    let ci = read(".github/workflows/ci.yml");
+    let release = read(".github/workflows/release.yml");
+    let makefile = read("Makefile");
+    let vscode_settings = read(".vscode/settings.json");
+    let vscode_launch = read(".vscode/launch.json");
+    let extension_readme = read("vscode-extension/README.md");
+    let extension_package = read("vscode-extension/package.json");
+    let extension_source = read("vscode-extension/src/extension.ts");
+    let setup_lsp = read("setup_lsp.sh");
+
+    assert!(ci.contains("cargo fmt --check"));
+    assert!(ci.contains("cargo clippy -- -D warnings"));
+    assert!(ci.contains("cargo test --lib"));
+    assert!(ci.contains("cargo test --tests"));
+    assert!(
+        ci.contains("types: [opened, reopened, ready_for_review]"),
+        "CI pull_request triggers must avoid synchronize spam while still running when PRs leave draft"
+    );
+    assert!(
+        !ci.contains("synchronize"),
+        "CI workflow should not run on every draft PR synchronize push"
+    );
+    assert_eq!(
+        ci.matches("github.event.pull_request.draft == false")
+            .count(),
+        3,
+        "CI fmt, clippy, and test jobs must keep the draft-PR guard"
+    );
+    assert!(
+        ci.contains("workflow_dispatch"),
+        "CI must keep a manual dispatch path when draft PR pushes do not run checks"
+    );
+
+    for unsupported in ["LLVM", "zen-lsp", "aarch64-apple-darwin"] {
+        assert!(
+            !release.contains(unsupported),
+            "release workflow advertises unsupported target/artifact: {unsupported}"
+        );
+        assert!(
+            !makefile.contains(unsupported),
+            "Makefile advertises unsupported target/artifact: {unsupported}"
+        );
+        assert!(
+            !vscode_settings.contains(unsupported),
+            ".vscode/settings.json advertises unsupported target/artifact: {unsupported}"
+        );
+        assert!(
+            !vscode_launch.contains(unsupported),
+            ".vscode/launch.json advertises unsupported target/artifact: {unsupported}"
+        );
+        assert!(
+            !extension_readme.contains(unsupported),
+            "vscode-extension/README.md advertises unsupported target/artifact: {unsupported}"
+        );
+        assert!(
+            !extension_package.contains(unsupported),
+            "vscode-extension/package.json advertises unsupported target/artifact: {unsupported}"
+        );
+        assert!(
+            !extension_source.contains(unsupported),
+            "vscode-extension/src/extension.ts advertises unsupported target/artifact: {unsupported}"
+        );
+        assert!(
+            !setup_lsp.contains(unsupported),
+            "setup_lsp.sh advertises unsupported target/artifact: {unsupported}"
+        );
+    }
+}
+
+#[test]
+fn checked_in_configs_do_not_contain_secret_literals() {
+    let config_paths = [
+        ".github/copilot-mcp.json",
+        ".github/workflows/ci.yml",
+        ".github/workflows/release.yml",
+        ".vscode/settings.json",
+        ".vscode/launch.json",
+        ".vscode/zen-language-config.json",
+    ];
+
+    for path in config_paths {
+        let contents = read(path);
+        for marker in ["Bearer ", "sk-", "ghp_", "github_pat_", "nk_"] {
+            assert!(
+                !contents.contains(marker),
+                "{path} contains credential-looking marker {marker}"
+            );
+        }
+    }
+}
