+++ tests/docs_truth/v1_spec.rs
@@ -1,0 +1,215 @@
+use super::*;
+
+#[test]
+fn contributor_docs_require_tests_first_for_language_work() {
+    let contributing = read("CONTRIBUTING.md");
+
+    for required in [
+        "Failing tests first",
+        "parser",
+        "semantic",
+        "effects",
+        "stdlib",
+        "codegen",
+        "tooling",
+        "C backend",
+    ] {
+        assert!(
+            contributing.contains(required),
+            "CONTRIBUTING.md is missing TDD/baseline requirement: {required}"
+        );
+    }
+
+    for stale_claim in [
+        "LLVM code generation",
+        "src/lsp",
+        "cargo build --bin zen-lsp",
+    ] {
+        assert!(
+            !contributing.contains(stale_claim),
+            "CONTRIBUTING.md still documents unsupported rewrite-baseline workflow: {stale_claim}"
+        );
+    }
+}
+
+#[test]
+fn v1_spec_records_phase_one_feature_gates_and_test_backlog() {
+    let spec = read("docs/V1_SPEC.md");
+
+    for required in [
+        "Status: v1 draft",
+        "Feature Matrix",
+        "implemented",
+        "gated",
+        "experimental",
+        "removed",
+        "Sync/Async effects",
+        "Typed allocators",
+        "Type matching",
+        "Behavior association",
+        "AST traversal",
+        "Actors in std",
+        "JSON/YAML IR boundaries",
+        "zen emit-json ast <file>",
+        "zen emit-json symbols <file>",
+        "zen emit-json typed <file>",
+        "zen emit-json diagnostics <file>",
+        "build.zen",
+        "zen test build.zen",
+        "test target execution",
+        "Executable target dependencies compile",
+        "dependency cycles",
+        "dependency-ordered multi-executable build tests",
+        "Accepted Syntax Forms",
+        "Test Evidence",
+        "Planned Positive Test",
+        "Planned Negative Test",
+        "resolver_records_value_symbol_generic_parameter_counts",
+        "resolver_records_value_symbol_function_type_metadata",
+        "resolver_records_value_symbol_generic_bounds",
+        "resolver_records_type_and_behavior_generic_parameter_counts",
+        "resolver_rejects_duplicate_type_parameter_names",
+        "resolver_records_generic_struct_field_types",
+        "resolver_records_generic_enum_variant_payload_types",
+        "resolver_records_generic_behavior_method_signatures",
+        "resolver_records_generic_behavior_function_type_method_signatures",
+        "parse_public_behavior_declaration",
+        "resolver_records_public_visibility_for_exported_declarations",
+        "resolver_rejects_duplicate_signature_parameter_names",
+        "check_program_with_symbols_validates_resolver_function_type_parameter_metadata",
+        "check_program_with_symbols_validates_resolver_function_type_return_metadata",
+        "check_program_with_symbols_validates_resolver_function_typed_signature_metadata",
+        "collect_declarations_with_symbols_uses_resolver_function_type_metadata",
+        "collect_declarations_with_symbols_uses_resolver_generic_function_template_metadata",
+        "collect_declarations_with_symbols_uses_resolver_generic_method_template_metadata",
+        "check_program_with_symbols_validates_resolver_function_type_parameter_names",
+        "check_program_with_symbols_validates_resolver_function_type_parameter_bound_refs",
+        "check_program_with_symbols_validates_resolver_type_parameter_names",
+        "behavior_impl_generic_behavior_type_arg_bound_passes_when_satisfied",
+        "behavior_impl_generic_behavior_type_arg_bound_failure_is_error",
+        "behavior_extends_generic_parent_accepts_child_type_parameter_arg",
+        "check_program_with_symbols_validates_resolver_generic_struct_field_types",
+        "check_program_with_symbols_validates_resolver_generic_enum_payload_types",
+        "check_program_with_symbols_validates_resolver_generic_behavior_method_signatures",
+        "check_program_with_symbols_validates_resolver_generic_behavior_function_type_method_signatures",
+        "check_program_with_symbols_validates_resolver_behavior_type_parameter_bounds",
+        "resolver_rejects_duplicate_struct_field_names",
+        "resolver_rejects_duplicate_struct_literal_fields",
+        "resolver_rejects_unknown_struct_literal_fields",
+        "resolver_rejects_missing_struct_literal_fields",
+        "resolver_rejects_unknown_struct_literal_types",
+        "resolver_allows_same_variant_names_in_different_enums",
+        "resolver_rejects_duplicate_variant_names_in_same_enum",
+        "resolver_rejects_unknown_enum_variant_expressions",
+        "resolver_rejects_missing_enum_variant_payload_expressions",
+        "resolver_rejects_unexpected_enum_variant_payload_expressions",
+        "tests/zen/duplicate_enum_variant_names.zen",
+        "check_module_graph_entry_seeds_imported_function_type_signatures",
+        "check_module_graph_entry_specializes_imported_generic_functions",
+        "check_module_graph_entry_specializes_imported_generic_enums",
+        "check_module_graph_entry_seeds_public_methods_for_imported_types",
+        "check_module_graph_entry_specializes_public_generic_methods_for_imported_types",
+        "tests/zen/multi_file_generic/main.zen",
+        "tests/zen/multi_file_generic_imported_type_dependency/main.zen",
+        "tests/zen/multi_file_generic_imported_transitive_dependency/main.zen",
+        "tests/zen/multi_file_type_method_worklist/main.zen",
+        "tests/zen/multi_file_type_method_method_dependency/main.zen",
+        "tests/zen/multi_file_type_method_imported_dependency/main.zen",
+        "tests/zen/multi_file_type_impl_imported_type_dependency/main.zen",
+        "tests/zen/multi_file_behavior_bound/main.zen",
+        "tests/zen/multi_file_behavior_inheritance/main.zen",
+        "tests/zen/multi_file_imported_behavior_impl/main.zen",
+        "tests/zen/multi_file_imported_behavior_default/main.zen",
+        "tests/zen/multi_file_imported_impl_imported_behavior/main.zen",
+        "tests/zen/multi_file_imported_child_parent_dispatch/main.zen",
+        "tests/zen/multi_file_imported_behavior_requires/main.zen",
+        "tests/zen/multi_file_imported_function_imported_behavior_bound/main.zen",
+        "tests/zen/multi_file_imported_function_param_type_dependency/main.zen",
+        "tests/zen/multi_file_imported_function_return_type_dependency/main.zen",
+        "tests/zen/multi_file_imported_function_imported_return_type_behavior/main.zen",
+        "tests/zen/multi_file_imported_generic_function_return_enum_dependency/main.zen",
+        "generated_c_call_definition_scan_reports_missing_generated_calls",
+        "generated_c_definition_count_ignores_prototypes",
+        "imported_function_signature_type_dependencies_are_not_directly_visible",
+        "imported_behavior_extends_requires_parent_methods",
+        "imported_behavior_extends_imported_parent_requires_parent_methods",
+        "imported_behavior_extends_requires_transitive_parent_methods",
+        "resolver_rejects_duplicate_behavior_impl_edges",
+        "resolver_rejects_duplicate_behavior_required_edges",
+        "resolver_rejects_duplicate_behavior_parent_edges",
+        "tests/zen/generic_method_worklist.zen",
+        "resolver_records_behavior_impl_methods_as_value_symbols",
+        "resolver_records_behavior_impl_function_type_methods",
+        "check_program_with_symbols_validates_resolver_impl_method_signature",
+        "check_program_with_symbols_validates_resolver_impl_function_type_signature",
+        "imported_private_behavior_impl_methods_are_not_directly_visible",
+        "resolver_rejects_duplicate_behavior_method_names",
+        "check_program_with_symbols_validates_resolver_generic_behavior_impl_names",
+        "check_program_with_symbols_validates_resolver_generic_behavior_required_names",
+        "check_program_with_symbols_validates_resolver_generic_behavior_parent_names",
+        "check_program_with_symbols_validates_resolver_generic_behavior_impl_refs",
+        "check_program_with_symbols_validates_resolver_generic_behavior_required_refs",
+        "check_program_with_symbols_validates_resolver_generic_behavior_parent_refs",
+        "collect_declarations_with_symbols_uses_resolver_behavior_impl_metadata",
+        "collect_declarations_with_symbols_uses_resolver_behavior_parent_metadata",
+        "resolver_records_method_signatures_as_value_symbols",
+        "resolver_records_method_function_type_signatures",
+        "check_program_with_symbols_validates_resolver_method_signature",
+        "check_program_with_symbols_validates_resolver_method_function_type_signature",
+        "check_module_graph_entry_does_not_seed_private_methods_for_imported_types",
+        "tests/zen/multi_file_type_method/main.zen",
+        "test_multi_file_type_method_worklist_imports",
+        "test_multi_file_type_method_method_dependency_imports",
+        "test_multi_file_type_method_imported_dependency_imports",
+        "imported_type_method_worklist_helpers_are_not_directly_visible",
+        "imported_type_method_dependencies_are_not_directly_visible",
+        "imported_type_method_imported_dependencies_are_not_directly_visible",
+        "imported_type_impl_imported_type_dependencies_are_not_directly_visible",
+        "imported_generic_function_imported_type_dependencies_are_not_directly_visible",
+        "imported_generic_function_transitive_dependencies_are_not_directly_visible",
+        "parse_impl_block",
+        "resolver_accepts_non_behavior_impl_blocks_as_method_symbols",
+        "resolver_rejects_duplicate_non_behavior_impl_method_names",
+        "resolver_rejects_non_behavior_impl_method_colliding_with_top_level_method",
+        "tests/zen/type_impl_methods.zen",
+        "tests/zen/multi_file_type_impl/main.zen",
+        "imported_private_type_impl_methods_are_not_visible",
+        "resolver_records_struct_function_type_fields",
+        "check_program_with_symbols_validates_resolver_struct_function_type_fields",
+        "check_program_with_symbols_validates_resolver_struct_typed_field_metadata",
+        "collect_declarations_with_symbols_uses_resolver_struct_field_metadata",
+        "resolver_records_enum_function_type_payloads",
+        "check_program_with_symbols_validates_resolver_enum_function_type_payloads",
+        "check_program_with_symbols_validates_resolver_enum_typed_payload_metadata",
+        "collect_declarations_with_symbols_uses_resolver_enum_payload_metadata",
+        "resolver_records_generic_enum_function_type_payloads",
+        "check_program_with_symbols_validates_resolver_generic_enum_function_type_payloads",
+        "resolver_records_behavior_function_type_method_signatures",
+        "check_program_with_symbols_validates_resolver_behavior_function_type_method_signatures",
+        "check_program_with_symbols_validates_resolver_behavior_method_types",
+        "collect_declarations_with_symbols_uses_resolver_behavior_method_metadata",
+        "resolver_records_generic_behavior_function_type_method_signatures",
+        "check_program_with_symbols_validates_resolver_generic_behavior_function_type_method_signatures",
+        "resolver_records_behavior_impl_method_body_locals",
+        "check_program_with_symbols_requires_resolver_impl_method_body_locals",
+        "resolver_records_behavior_default_method_body_locals",
+        "check_program_with_symbols_requires_resolver_behavior_default_locals",
+        "resolver_records_struct_field_default_locals",
+        "check_program_with_symbols_requires_resolver_struct_field_default_locals",
+        "resolver_records_top_level_expr_locals",
+        "check_program_with_symbols_requires_resolver_top_level_expr_locals",
+        "resolver_records_closure_locals",
+        "check_program_with_symbols_requires_resolver_closure_locals",
+        "resolver_records_mutable_closure_parameter_locals",
+        "check_program_with_symbols_validates_resolver_closure_parameter_mutability",
+        "resolver_records_pattern_locals",
+        "check_program_with_symbols_requires_resolver_pattern_locals",
+        "resolver_records_same_name_locals_in_distinct_scopes",
+        "check_program_with_symbols_validates_resolver_local_mutability_by_scope",
+    ] {
+        assert!(
+            spec.contains(required),
+            "docs/V1_SPEC.md is missing Phase 1 requirement: {required}"
+        );
+    }
+}
